changed type for budget

+++ parser-alert-job/src/main/java/by/gdev/alert/job/parser/domain/freelancer/BudgetFreelancerOrder.java
@@ -1,11 +1,12 @@
 package by.gdev.alert.job.parser.domain.freelancer;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+
 import lombok.Data;
 
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class BudgetFreelancerOrder {
-    private Integer minimum;
-    private Integer maximum;
+	private Long minimum;
+	private Long maximum;
 }
